Add one more API to allow impersonation request from privileged process
This commit supports one more use case for Hive and Presto users where doAs is turned off hence the container is running as Hive and Presto user. To retrieve data from s3, we are going to make changes in EMRFS so that EMRFS directly calls the new API to get credentials (impersonation) for the actual end user.

We are going to validate that the caller of this API is pre-approved according to the configuration.